fix/에너지바
에너지바 로직을
기존 : int에서, 시간이 지나면 int 숫자가 다 차는 식
변경 : float로 변경 후, 표시되는 숫자만 소숫점을 뗀 식으로

+++ Assets/Scripts/Base/GameRule/GameCondition.cs
@@ -20,7 +20,7 @@ public class GameCondition : MonoBehaviour
         P_AP = Player.GetComponent<PlayerStatus>().playerAP;
         P_AR = Player.GetComponent<PlayerStatus>().playerAR;
         P_MR = Player.GetComponent<PlayerStatus>().playerMR;
-        Cur_Cost = UI_EnergyBar.Instance.currentHealth;
+        Cur_Cost = (int)UI_EnergyBar.Instance.currentHealth;
         Cost_Fill = UI_EnergyBar.Instance.fillSpeed;
         
         foreach(var relic in RelicManager.Instance.activeRelicsDict)
@@ -42,7 +42,7 @@ public class GameCondition : MonoBehaviour
         P_AP = Player.GetComponent<PlayerStatus>().playerAP;
         P_AR = Player.GetComponent<PlayerStatus>().playerAR;
         P_MR = Player.GetComponent<PlayerStatus>().playerMR;
-        Cur_Cost = UI_EnergyBar.Instance.currentHealth;
+        Cur_Cost = (int)UI_EnergyBar.Instance.currentHealth;
         Cost_Fill = UI_EnergyBar.Instance.fillSpeed;
 
         RelicManager.Instance.CheckRelics(this);
